pyproject.toml
[tool.poetry]
name = "reflex-chat"
version = "0.1.0"
description = ""
authors = ["Nicolas"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
azure-core = "1.32.0"
langchain-community = "0.3.14"
reflex = "0.6.7"
openai = "1.59.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
